+++ packages/jest/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @vslint/jest
 
+## 0.3.1
+
+### Patch Changes
+
+- 9fc3298: Tiny bump on all
+- Updated dependencies [9fc3298]
+  - @vslint/shared@0.0.12
+
+## 0.3.0
+
+### Minor Changes
+
+- fefbde7: Using jest snapshots instead of custom implementation
+
+### Patch Changes
+
+- Updated dependencies [fefbde7]
+  - @vslint/shared@0.0.11
+
 ## 0.2.16
 
 ### Patch Changes