+++ packages/server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @vslint/server
 
+## 0.2.9
+
+### Patch Changes
+
+- 9fc3298: Tiny bump on all
+- Updated dependencies [9fc3298]
+  - @vslint/shared@0.0.12
+
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [fefbde7]
+  - @vslint/shared@0.0.11
+
 ## 0.2.7
 
 ### Patch Changes