+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vslint/shared
 
+## 0.0.12
+
+### Patch Changes
+
+- 9fc3298: Tiny bump on all
+
+## 0.0.11
+
+### Patch Changes
+
+- fefbde7: Using jest snapshots instead of custom implementation
+
 ## 0.0.10
 
 ### Patch Changes